Full project

+++ src/main/java/service/ProductServiceImpl.java
@@ -11,25 +11,40 @@ public class ProductServiceImpl implements ProductService {
     private final ProductDao productDao = new ProductDao();
 
     public void saveProduct(Product product) {
-        isValidProduct(product);
-        productDao.save(product);
+        if (isValidProduct(product) && isExistProduct(product)) {
+            productDao.save(product);
+        }
     }
 
     public void updateProduct(Product product) {
-        isValidProduct(product);
-        productDao.update(product);
+        if (isValidProduct(product)) {
+            productDao.update(product);
+        }
     }
 
     public void removeProduct(Product product) {
-        isValidProduct(product);
-        productDao.delete(product);
+        if (isValidProduct(product)) {
+            productDao.delete(product);
+        }
     }
 
-    private void isValidProduct(Product product) {
+    private boolean isValidProduct(Product product) {
         boolean isValid = product.getPrice() > 0 && product.getAmount() >= 0 && product.getCategory() != null;
         if (!isValid) {
             System.out.println("Product data is not valid.");
+            return false;
         }
+        return true;
+    }
+
+    private boolean isExistProduct(Product product) {
+        for (Product tempProduct : productDao.getAll()) {
+            if (tempProduct.getName().equals(product.getName()) && tempProduct.getCategory().equals(product.getCategory())) {
+                System.out.println("Product data is exist");
+                return false;
+            }
+        }
+        return true;
     }
 
     public List<Product> getAllProducts() {
@@ -42,5 +57,9 @@ public class ProductServiceImpl implements ProductService {
 
     public List<Product> searchProductsByCategory(ProductCategories category) {
         return productDao.getByCategory(category);
+    }
+
+    public Product getProductById(String Id) {
+        return productDao.getById(Id);
     }
 }
